+++ src/api/users/handlers/delete.ts
@@ -5,6 +5,11 @@ import prisma from "../../../../prisma/client";
 const deleteOneUser: IUserHandlers["delete"] = async (req, res) => {
   try {
     const { id } = req.params;
+    if (!req.user || req.user.role !== "SUPER_ADMIN") {
+      return res
+        .status(403)
+        .json({ message: "Forbidden, you don't have the right access" });
+    }
     const deletedUser = await prisma.user.delete({
       where: {
         id,
